move units dict to pandas utils
